+++ errors.go
@@ -266,7 +266,7 @@ func (f *fundamental) Format(s fmt.State, verb rune) {
 					io.WriteString(s, "\n")
 				}
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			f.stack.Format(s, verb)
 			return
 		}
@@ -309,7 +309,7 @@ func (w *errorf) Format(s fmt.State, verb rune) {
 					io.WriteString(s, "\n")
 				}
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			stack(w.StackTrace()).Format(s, verb)
 			return
 		}
@@ -365,7 +365,7 @@ func (w *errorfWithStack) Format(s fmt.State, verb rune) {
 					io.WriteString(s, "\n")
 				}
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			w.stack.Format(s, verb)
 			return
 		}
@@ -436,7 +436,7 @@ func (w *withStack) Format(s fmt.State, verb rune) {
 					io.WriteString(s, "\n")
 				}
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			w.stack.Format(s, verb)
 			return
 		}
@@ -569,7 +569,7 @@ func (w *wrapped) Format(s fmt.State, verb rune) {
 					io.WriteString(s, "\n")
 				}
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			w.stack.Format(s, verb)
 			unwrap := fmt.Sprintf("%+v", w.error)
 			if len(unwrap) > 0 {
@@ -791,7 +791,7 @@ func (w *withMessageAndStack) Format(s fmt.State, verb rune) {
 			if needsNewline {
 				io.WriteString(s, "\n")
 			}
-			fmt.Fprintf(s, "Stack trace (most recent call first):\n")
+			fmt.Fprintf(s, "stack trace (most recent call first):\n")
 			w.stack.Format(s, verb)
 			return
 		}