+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2023-06-06
+
 ### Added
 
 - `StackFormat` to format the provided stack trace as text.
@@ -57,7 +59,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.4.1...main
+[unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.5.0...main
+[0.5.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.4.1...v0.5.0
 [0.4.1]: https://gitlab.com/tozd/go/errors/-/compare/v0.4.0...v0.4.1
 [0.4.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.3.0...v0.4.0
 [0.3.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.2.0...v0.3.0